adds helper methods to helpers.py
