+++ tests/opentrons/labware/test_pipette.py
@@ -363,197 +363,6 @@ class PipetteTest(unittest.TestCase):
         self.p200.set_speed(dispense=100)
         self.assertEqual(self.p200.speeds['dispense'], 100)
 
-    def test_create_well_pairs(self):
-
-        s = self.plate[0]
-        t = self.plate[1]
-        result = self.p200._create_well_pairs(s, t)
-        expected = ([s], [t])
-        self.assertEquals(expected, result)
-
-        s = self.plate[0:3]
-        t = self.plate[1]
-        result = self.p200._create_well_pairs(s, t)
-        expected = (s, [t] * len(s))
-        self.assertEquals(expected, result)
-
-        s = self.plate[0]
-        t = self.plate[0:3]
-        result = self.p200._create_well_pairs(s, t)
-        expected = ([s] * len(t), t)
-        self.assertEquals(expected, result)
-
-        s = self.plate[0:3]
-        t = self.plate[0:3]
-        result = self.p200._create_well_pairs(s, t)
-        expected = (s, t)
-        self.assertEquals(expected, result)
-
-        s = self.plate.rows[0:3]
-        t = self.plate[0]
-        result = self.p200._create_well_pairs(s, t)
-        expected = (s, [t] * 3)
-        self.assertEquals(expected, result)
-
-        s = self.plate.rows[0:3]
-        t = self.plate[0:3]
-        result = self.p200._create_well_pairs(s, t)
-        expected = (s, t)
-        self.assertEquals(expected, result)
-
-        s = self.plate[0:3]
-        t = self.plate.rows[0:3]
-        result = self.p200._create_well_pairs(s, t)
-        expected = (s, t)
-        self.assertEquals(expected, result)
-
-        s = self.plate[0:3]
-        t = self.plate.rows[0:2]
-        self.assertRaises(RuntimeError, self.p200._create_well_pairs, s, t)
-
-    def test_create_volume_pairs(self):
-        result = self.p200._create_volume_pairs(20, 1)
-        expected = [20]
-        self.assertEquals(expected, result)
-
-        result = self.p200._create_volume_pairs([20], 1)
-        expected = [20]
-        self.assertEquals(expected, result)
-
-        self.assertRaises(
-            RuntimeError, self.p200._create_volume_pairs, [20, 2], 1)
-
-        result = self.p200._create_volume_pairs([0, 10], 11)
-        expected = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-        self.assertEquals(expected, result)
-
-        result = self.p200._create_volume_pairs(
-            [0, 10], 11, interpolate=lambda x: 1 - x)
-        expected = [10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0]
-        for i in range(len(result)):
-            self.assertAlmostEquals(result[i], expected[i])
-
-    def test_match_volumes_to_sources(self):
-        s = [self.plate[0], self.plate[0], self.plate[0]]
-        v = [1, 1, 2]
-        result = self.p200._match_volumes_to_sources(v, s)
-        expected = [1, 1, 2]
-        self.assertEquals(result, expected)
-
-        s = [self.plate[0], self.plate[0], self.plate[1]]
-        v = [1, 1, 2]
-        result = self.p200._match_volumes_to_sources(v, s)
-        expected = [1, 1]
-        self.assertEquals(result, expected)
-
-        s = [self.plate[0], self.plate[1], self.plate[0]]
-        v = [1, 1, 2]
-        result = self.p200._match_volumes_to_sources(v, s)
-        expected = [1]
-        self.assertEquals(result, expected)
-
-    def test_find_aspirate_volume(self):
-
-        result = self.p200._find_aspirate_volume([0])
-        self.assertEquals(result, 0)
-
-        result = self.p200._find_aspirate_volume(
-            [0, 1, 2, 3, 4, 5])
-        self.assertEquals(result, 15)
-
-        result = self.p200._find_aspirate_volume(
-            [0, 1, 2, 3, 4, 5, 186])
-        self.assertEquals(result, 15)
-
-        result = self.p200._find_aspirate_volume(
-            [0, 1, 2, 3, 4, 5, 185])
-        self.assertEquals(result, 200)
-
-    def test_transfer_single(self):
-
-        self.p200.reset()
-        self.p200._transfer_single(
-            [200],
-            self.plate[0:1],
-            self.plate[0:1]
-        )
-        expected = [
-            ['aspirating', '200'],
-            ['touch'],
-            ['dispensing', '200'],
-            ['touch'],
-            ['blow']
-        ]
-        self.assertEqual(len(self.robot.commands()), len(expected))
-        for i, c in enumerate(self.robot.commands()):
-            for s in expected[i]:
-                self.assertTrue(s.lower() in c.lower())
-        self.robot.clear_commands()
-
-        self.p200.reset()
-        self.p200._transfer_single(
-            [10] * 8,
-            [self.plate[0]] * 8,
-            self.plate[1:9]
-        )
-        expected = [
-            ['aspirating', '80'],
-            ['touch'],
-            ['dispensing', '10'],
-            ['touch']
-        ]
-        self.assertEqual(len(self.robot.commands()), len(expected))
-        for i, c in enumerate(self.robot.commands()):
-            for s in expected[i]:
-                self.assertTrue(s.lower() in c.lower())
-        self.robot.clear_commands()
-
-        self.p200.reset()
-        self.p200._transfer_single(
-            [10] * 8,
-            self.plate[0:8],
-            self.plate[1:9]
-        )
-        expected = [
-            ['aspirating', '10'],
-            ['touch'],
-            ['dispensing', '10'],
-            ['touch'],
-            ['blow']
-        ]
-        self.assertEqual(len(self.robot.commands()), len(expected))
-        for i, c in enumerate(self.robot.commands()):
-            for s in expected[i]:
-                self.assertTrue(s.lower() in c.lower())
-        self.robot.clear_commands()
-
-        self.p200.reset()
-        self.p200._transfer_single(
-            [10] * 8,
-            self.plate[0:8],
-            self.plate[1:9],
-            mix=(3, 10)
-        )
-        expected = [
-            ['aspirating', '10'],
-            ['touch'],
-            ['dispensing', '10'],
-            ['mix', '10'],
-            ['aspirating', '10'],
-            ['dispensing', '10'],
-            ['aspirating', '10'],
-            ['dispensing', '10'],
-            ['aspirating', '10'],
-            ['dispensing', '10'],
-            ['touch'],
-            ['blow']
-        ]
-        self.assertEqual(len(self.robot.commands()), len(expected))
-        for i, c in enumerate(self.robot.commands()):
-            for s in expected[i]:
-                self.assertTrue(s.lower() in c.lower())
-        self.robot.clear_commands()
-
     def test_transfer(self):
         self.p200.reset()
         self.p200.transfer(
